refactor: using slog function to show log messages
