+++ image_hosting/api/uploadImage.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"context"
 	"github.com/gofiber/fiber/v2"
 	"github.com/opentreehole/backend/common"
 	. "github.com/opentreehole/backend/image_hosting/config"
@@ -8,51 +9,53 @@ import (
 	. "github.com/opentreehole/backend/image_hosting/schema"
 	. "github.com/opentreehole/backend/image_hosting/utils"
 	"io"
-	"log"
+	"log/slog"
 	"path/filepath"
 	"strings"
 	"time"
 )
 
 func UploadImage(c *fiber.Ctx) error {
-	log.Println("uploading image")
+	slog.LogAttrs(context.Background(), slog.LevelInfo, "uploading image")
+
 	// response to frontend
 	var response CheveretoUploadResponse
 	// the file uploaded by user  in the request body with the form-data key "source"
 	file, err := c.FormFile("source")
 	if err != nil {
-		log.Println(err)
+		slog.LogAttrs(context.Background(), slog.LevelError, "No file uploaded", slog.String("err", err.Error()))
 		return common.BadRequest("No file uploaded")
 	}
 
 	fileSize := file.Size
 	maxSize := 10 * 1024 * 1024 // file should <= 10MB
 	if int(fileSize) > maxSize {
+		slog.LogAttrs(context.Background(), slog.LevelError, "File size is too large")
 		return common.BadRequest("File size is too large")
 	}
 
 	fileExtension := strings.TrimPrefix(filepath.Ext(file.Filename), ".")
 
 	if !IsAllowedExtension(fileExtension) {
-		log.Println("Error: File type not allowed.")
+		slog.LogAttrs(context.Background(), slog.LevelError, "File type not allowed.")
 		return common.BadRequest("File type not allowed.")
 	}
 
 	fileContent, err := file.Open()
 	if err != nil {
-		log.Println(err)
+		slog.LogAttrs(context.Background(), slog.LevelError, "The uploaded file has some problems", slog.String("err", err.Error()))
 		return common.BadRequest("The uploaded file has some problems")
 	}
 
 	imageData, err := io.ReadAll(fileContent)
 	if err != nil {
-		log.Println(err)
+		slog.LogAttrs(context.Background(), slog.LevelError, "The uploaded file has some problems", slog.String("err", err.Error()))
 		return common.BadRequest("The uploaded file has some problems")
 	}
 
 	imageIdentifier, err := GenerateIdentifier()
 	if err != nil {
-		log.Println(err)
+		slog.LogAttrs(context.Background(), slog.LevelError, "Cannot generate image identifier", slog.String("err", err.Error()))
 		return common.InternalServerError("Cannot generate image identifier")
 	}
 
@@ -65,7 +68,7 @@ func UploadImage(c *fiber.Ctx) error {
 	err = DB.Create(&uploadedImage).Error
 
 	if err != nil {
-		log.Println(err)
+		slog.LogAttrs(context.Background(), slog.LevelError, "Database cannot store the image", slog.String("err", err.Error()))
 		return common.InternalServerError("Database cannot store the image")
 	}
 
@@ -80,26 +83,7 @@ func UploadImage(c *fiber.Ctx) error {
 		DisplayUrl: imageUrl,
 		Mime:       "image/" + fileExtension,
 	}
-	log.Printf("image is uploaded on: %v\n", imageUrl)
+	slog.LogAttrs(context.Background(), slog.LevelInfo, "Image uploaded", slog.String("url", imageUrl))
 	return c.JSON(&response)
 
-}
-
-func GetImage(c *fiber.Ctx) error {
-	// to access the image in database
-	// year := c.Params("year")
-	// month := c.Params("month")
-	// day := c.Params("day")
-	// imageType := strings.Split(c.Params("imageIdentifier"), ".")[1]
-	log.Println("getting image")
-	var image ImageTable
-	imageIdentifier := strings.Split(c.Params("identifier"), ".")[0]
-	err := DB.First(&image, "base_name = ?", imageIdentifier)
-	if err.Error != nil {
-		log.Println(err)
-		return common.BadRequest("Cannot find the image")
-	}
-	log.Println("get image successfully")
-	// browser will automatically transform the BLOB data to an image (no matter what extension)
-	return c.Send(image.ImageFileData)
 }
